requirements.txt
Django==1.10.4
django-bootstrap3==7.1.0
django-watson==1.2.4
Pillow==3.4.2
wheel==0.24.0
